+++ src/modules/cast/cast.cu
@@ -1,6 +1,7 @@
 #include <cuda_fp16.h>
+#include <stdint.h>
 
-template<typename IT, typename OT, size_t N>
+template<typename IT, typename OT, uint64_t N>
 __global__ void cast(IT* input, OT* output) {
     const int tid = blockIdx.x * blockDim.x + threadIdx.x;
